// apps/backend/src/integrations/strava/types.gen.ts
export type LatLng = [number, number];
// ...
export type AltitudeStream = BaseStream & {
  /**
   * The sequence of altitude values for this stream, in meters
   */
  data?: Array<number>;
};

export type BaseStream = {
  /**
   * The number of data points in this stream
   */
  original_size?: number;
  /**
   * The level of detail (sampling) in which this stream was returned
   */
  resolution?: "low" | "medium" | "high";
  /**
   * The base series used in the case the stream was downsampled
   */
  series_type?: "distance" | "time";
};

export type CadenceStream = BaseStream & {
  /**
   * The sequence of cadence values for this stream, in rotations per minute
   */
  data?: Array<number>;
};

export type DistanceStream = BaseStream & {
  /**
   * The sequence of distance values for this stream, in meters
   */
  data?: Array<number>;
};

export type HeartrateStream = BaseStream & {
  /**
   * The sequence of heart rate values for this stream, in beats per minute
   */
  data?: Array<number>;
};

export type LatLngStream = BaseStream & {
  /**
   * The sequence of lat/long values for this stream
   */
  data?: Array<LatLng>;
};

export type MovingStream = BaseStream & {
  /**
   * The sequence of moving values for this stream, as boolean values
   */
  data?: Array<boolean>;
};

export type PowerStream = BaseStream & {
  /**
   * The sequence of power values for this stream, in watts
   */
  data?: Array<number>;
};

export type SmoothGradeStream = BaseStream & {
  /**
   * The sequence of grade values for this stream, as percents of a grade
   */
  data?: Array<number>;
};

export type SmoothVelocityStream = BaseStream & {
  /**
   * The sequence of velocity values for this stream, in meters per second
   */
  data?: Array<number>;
};

export type StreamSet = {
  time?: TimeStream;
  distance?: DistanceStream;
  latlng?: LatLngStream;
  altitude?: AltitudeStream;
  velocity_smooth?: SmoothVelocityStream;
  heartrate?: HeartrateStream;
  cadence?: CadenceStream;
  watts?: PowerStream;
  temp?: TemperatureStream;
  moving?: MovingStream;
  grade_smooth?: SmoothGradeStream;
};

export type TemperatureStream = BaseStream & {
  /**
   * The sequence of temperature values for this stream, in celsius degrees
   */
  data?: Array<number>;
};

export type TimeStream = BaseStream & {
  /**
   * The sequence of time values for this stream, in seconds
   */
  data?: Array<number>;
};